fixed last

+++ server/api/controller/Record.py
@@ -1,4 +1,5 @@
 from api.models import Record, Diary
+from account.models import User
 from api.serializer import RecordSelializer
 
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
@@ -55,15 +56,26 @@ def GetRecord(id) -> RecordSelializer:
     except ObjectDoesNotExist:
         raise NotFound
 
-def DeleteRecord(id):
+def DeleteRecord(user_id, record_id):
     try:
-        record = Record.objects.filter(_id=ObjectId(id))
-        print(record)
-        record.delete()
+        user = User.objects.get(_id=ObjectId(user_id))
+        record = Record.objects.filter(_id=ObjectId(record_id))
+        if record.first() == None:
+            return None
+            
+        diaries = user.diaries.all()
+        exsits = False
+        for diary in diaries:
+            for _record in diary.records.all():
+                if _record._id == record.first()._id:
+                    exsits = True
+
+        if exsits:
+            record.delete()
+            serializer = RecordSelializer(record.first(), many=False)
+            return serializer.data
+        return None
         
-        serializer = RecordSelializer(record.first(), many=False)
-        # record = Record.objects.filter(pk=id).first()
-        return serializer.data
     except ObjectDoesNotExist:
         raise NotFound
